Repository setup and added for the project
