+++ Presentation/Web.config
@@ -16,7 +16,7 @@
   </appSettings>
   <connectionStrings>
     <!--<add name="StringDBContext" connectionString="Server=.;Initial Catalog=LGADB;Persist Security Info=False;Integrated Security=SSPI;MultipleActiveResultSets=True;Encrypt=False;TrustServerCertificate=False;Connection Timeout=30;" providerName="System.Data.SqlClient" />-->    
-    <add name="StringDBContext" connectionString="Server=(localdb)\MSSQLLocalDB;Initial Catalog=LGADB-Dev;Persist Security Info=False;Integrated Security=SSPI;MultipleActiveResultSets=True;Encrypt=False;TrustServerCertificate=False;Connection Timeout=30;" providerName="System.Data.SqlClient" />  
+    <add name="StringDBContext" connectionString="Server=(localdb)\MSSQLLocalDB;Initial Catalog=LGADB-Dev1;Persist Security Info=False;Integrated Security=SSPI;MultipleActiveResultSets=True;Encrypt=False;TrustServerCertificate=False;Connection Timeout=30;" providerName="System.Data.SqlClient" />  
   
   </connectionStrings>
   <system.web>
